Create evenXandO
Write a function that checks a string to see if it has the same amount of xs and os. The case of the characters should be ignored, so x and X are equivalent, as are o and O. The method must return a boolean. The string can contains any other char, which will be ignored.