Create FindProcesses
You will write a function which will be provided with a set of processes that each convert from one item to another.
These will be provided as an array, with each process in the form of "process_name:from_item:to_item". For example, you might have a process named "smelt" that converts ore into a metal, like this: "smelt:ore:metal".
You will also be passed in the initial start item and desired end item for the overall workflow.
Note: For simplicity, all processes and items will be simple alphanumeric strings.